+++ src/addon/cc-addon-option-form.js
@@ -4,6 +4,7 @@ import '../molecules/cc-block.js';
 import { css, html, LitElement } from 'lit-element';
 import { dispatchCustomEvent } from '../lib/events.js';
 import { i18n } from '../lib/i18n.js';
+import { defaultThemeStyles } from '../styles/default-theme.js';
 import { linkStyles } from '../templates/cc-link.js';
 
 /**
@@ -84,6 +85,7 @@ export class CcAddonOptionForm extends LitElement {
 
   static get styles () {
     return [
+      defaultThemeStyles,
       linkStyles,
       // language=CSS
       css`