+++ src/addon/cc-jenkins-info.js
@@ -5,6 +5,7 @@ import '../molecules/cc-error.js';
 import { css, html, LitElement } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { i18n } from '../lib/i18n.js';
+import { defaultThemeStyles } from '../styles/default-theme.js';
 import { skeletonStyles } from '../styles/skeleton.js';
 import { ccLink, linkStyles } from '../templates/cc-link.js';
 
@@ -103,8 +104,9 @@ export class CcJenkinsInfo extends LitElement {
 
   static get styles () {
     return [
-      skeletonStyles,
+      defaultThemeStyles,
       linkStyles,
+      skeletonStyles,
       // language=CSS
       css`
         :host {