+++ src/invoices/cc-invoice.js
@@ -5,6 +5,7 @@ import '../atoms/cc-html-frame.js';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { i18n } from '../lib/i18n.js';
+import { defaultThemeStyles } from '../styles/default-theme.js';
 import { skeletonStyles } from '../styles/skeleton.js';
 import { ccLink, linkStyles } from '../templates/cc-link.js';
 
@@ -82,6 +83,7 @@ export class CcInvoice extends LitElement {
 
   static get styles () {
     return [
+      defaultThemeStyles,
       linkStyles,
       skeletonStyles,
       // language=CSS