+++ stories/templates/cc-link.stories.js
@@ -1,4 +1,5 @@
 import { render } from 'lit-html';
+import { defaultThemeStyles } from '../../src/styles/default-theme.js';
 import { skeletonStyles } from '../../src/styles/skeleton.js';
 import { ccLink, linkStyles } from '../../src/templates/cc-link.js';
 import { makeStory } from '../lib/make-story.js';
@@ -10,7 +11,7 @@ export default {
 };
 
 const conf = {
-  css: [linkStyles, skeletonStyles].join(''),
+  css: [defaultThemeStyles, linkStyles, skeletonStyles].join(''),
 };
 
 export const defaultStory = makeStory(conf, {